# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Api",
    "version": "0.0.1"
  },
  "paths": {
...
    "/studio/save-assay-json": {
      "post": {
        "tags": ["studio"],
        "operationId": "studio.saveAssayJson",
        "parameters": [],
        "security": [],
        "responses": {
          "200": {
            "description": "SaveAssayJsonResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveAssayJsonResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request did not match the expected schema",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/HttpApiDecodeError"
                    },
                    {
                      "$ref": "#/components/schemas/RequestError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveAssayJsonRequest"
              }
            }
          },
          "required": true
        }
      }
    },
    "/studio/save-result-pdf": {
      "post": {
        "tags": ["studio"],
        "operationId": "studio.saveResultPdf",
        "parameters": [],
        "security": [],
        "responses": {
          "200": {
            "description": "SaveResultPdfResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveResultPdfResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request did not match the expected schema",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/HttpApiDecodeError"
                    },
                    {
                      "$ref": "#/components/schemas/RequestError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveResultPdfRequest"
              }
            }
          },
          "required": true
        }
      }
    },
    "/studio/start-analysis": {
      "post": {
        "tags": ["studio"],
        "operationId": "studio.startAnalysis",
        "parameters": [],
        "security": [],
        "responses": {
          "200": {
            "description": "AnalysisProgress",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisProgress"
                }
              }
            }
          },
          "400": {
            "description": "The request did not match the expected schema",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/HttpApiDecodeError"
                    },
                    {
                      "$ref": "#/components/schemas/RequestError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisStartRequest"
              }
            }
          },
          "required": true
        }
      }
    },
    "/studio/analysis-progress": {
      "get": {
        "tags": ["studio"],
        "operationId": "studio.getAnalysisProgress",
        "parameters": [],
        "security": [],
        "responses": {
          "200": {
            "description": "AnalysisProgress",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisProgress"
                }
              }
            }
          },
          "400": {
            "description": "The request did not match the expected schema",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/HttpApiDecodeError"
                    },
                    {
                      "$ref": "#/components/schemas/RequestError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/studio/latest-analysis": {
      "get": {
        "tags": ["studio"],
        "operationId": "studio.getLatestAnalysisProgress",
        "parameters": [],
        "security": [],
        "responses": {
          "200": {
            "description": "AnalysisProgress",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AnalysisProgress"
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request did not match the expected schema",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/HttpApiDecodeError"
                    },
                    {
                      "$ref": "#/components/schemas/RequestError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/studio/analysis-results": {
      "get": {
        "tags": ["studio"],
        "operationId": "studio.getAnalysisResults",
        "parameters": [],
        "security": [],
        "responses": {
          "200": {
            "description": "AnalysisProgress",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AnalysisProgress"
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request did not match the expected schema",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/HttpApiDecodeError"
                    },
                    {
                      "$ref": "#/components/schemas/RequestError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HttpApiDecodeError": {
        "type": "object",
        "required": ["issues", "message", "_tag"],
        "properties": {
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Issue"
            }
          },
          "message": {
            "type": "string"
          },
          "_tag": {
            "type": "string",
            "enum": ["HttpApiDecodeError"]
          }
        },
        "additionalProperties": false,
        "description": "The request did not match the expected schema"
      },
      "Issue": {
        "type": "object",
        "required": ["_tag", "path", "message"],
        "properties": {
          "_tag": {
            "type": "string",
            "enum": [
              "Pointer",
              "Unexpected",
              "Missing",
              "Composite",
              "Refinement",
              "Transformation",
              "Type",
              "Forbidden"
            ],
            "description": "The tag identifying the type of parse issue"
          },
          "path": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyKey"
            },
            "description": "The path to the property where the issue occurred"
          },
          "message": {
            "type": "string",
            "description": "A descriptive message explaining the issue"
          }
        },
        "additionalProperties": false,
        "description": "Represents an error encountered while parsing a value to match the schema"
      },
      "PropertyKey": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "number"
          },
          {
            "type": "object",
            "required": ["_tag", "key"],
            "properties": {
              "_tag": {
                "type": "string",
                "enum": ["symbol"]
              },
              "key": {
                "type": "string"
              }
            },
            "additionalProperties": false,
            "description": "an object to be decoded into a globally shared symbol"
          }
        ]
      },
      "RequestError": {
        "type": "object",
        "required": ["message", "_tag"],
        "properties": {
          "message": {
            "type": "string"
          },
          "_tag": {
            "type": "string",
            "enum": ["RequestError"]
          }
        },
        "additionalProperties": false
      },
      "Unauthorized": {
        "type": "object",
        "required": ["message", "_tag"],
        "properties": {
          "message": {
            "type": "string"
          },
          "_tag": {
            "type": "string",
            "enum": ["Unauthorized"]
          }
        },
        "additionalProperties": false
      },
...
      "SaveAssayJsonRequest": {
        "type": "object",
        "required": ["saveTo", "contents"],
        "properties": {
          "saveTo": {
            "type": "string"
          },
          "contents": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "SaveAssayJsonResponse": {
        "type": "object",
        "required": ["ok", "path"],
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "path": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "SaveResultPdfRequest": {
        "type": "object",
        "required": ["workspacePath", "fileName", "contentsBase64"],
        "properties": {
          "workspacePath": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "contentsBase64": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "SaveResultPdfResponse": {
        "type": "object",
        "required": ["ok", "directory", "path"],
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "directory": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "AnalysisStartRequest": {
        "type": "object",
        "required": ["workspacePath", "requestId"],
        "properties": {
          "workspacePath": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "AnalysisProgress": {
        "type": "object",
        "required": ["requestId", "status", "stage", "progress", "message", "error"],
        "properties": {
          "requestId": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatus"
          },
          "stage": {
            "$ref": "#/components/schemas/AnalysisStage"
          },
          "progress": {
            "type": "number",
            "format": "double"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "resultFiles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnalysisCsvFile"
            }
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false
      },
      "AnalysisStatus": {
        "type": "string",
        "enum": ["queued", "running", "completed", "error"]
      },
      "AnalysisStage": {
        "type": "string",
        "enum": ["queued", "preparing", "segment", "timeseries", "auc", "fit", "completed"]
      },
      "AnalysisCsvFile": {
        "type": "object",
        "required": ["kind", "fileName", "path", "csv"],
        "properties": {
          "kind": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "csv": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {}
  },
  "security": [],
...
}